+++ backend/ng/challenge/routes/__init__.py
@@ -1,5 +1,7 @@
 from .admin_routes import challenge_admin_namespace
+from .user_routes import challenge_namespace
 
 __all__ = [
     "challenge_admin_namespace",
+    "challenge_namespace"
 ]
